pyproject.toml
[project]
name = "baynext-ml"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-meridian>=1.1.4",
    "mlflow-skinny>=3.2.0",
    "psutil>=7.0.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "vl-convert-python>=1.8.0",
]


[project.scripts]
baynext = "baynext.cli:cli"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
]

[tool.uv]
package = true
